Introduce an initial directive in app.js, which allows me to update the fillcolor attribute (which appears to actually be a property on the dom element).
The plan is to extend this further, first to implement some simple
structures for VML with a few attributes/properties that can be
interpolated, and then to extend this to SVG as well, with the
directive being automatically set up 'correctly' for the environment
on startup. Then to try to pull in more and more SVG/VML elements as
appropriate, to provide a Raphael-style layer over the common
denominators of both, fit as much as possible into angular directives
as an idiom.